move non-security xDS-enabled server tests to their own file
