+++ test/cpp/end2end/xds/BUILD
@@ -593,3 +593,26 @@ grpc_cc_test(
         "@envoy_api//envoy/extensions/transport_sockets/http_11_proxy/v3:pkg_cc_proto",
     ],
 )
+
+grpc_cc_test(
+    name = "xds_enabled_server_end2end_test",
+    size = "large",
+    srcs = ["xds_enabled_server_end2end_test.cc"],
+    external_deps = [
+        "gtest",
+    ],
+    flaky = True,  # TODO(b/144705388)
+    linkstatic = True,  # Fixes dyld error on MacOS
+    tags = [
+        "no_test_ios",
+        "no_windows",
+        "xds_end2end_test",
+    ],  # TODO(jtattermusch): fix test on windows
+    deps = [
+        ":xds_end2end_test_lib",
+        "//:gpr",
+        "//:grpc",
+        "//:grpc++",
+        "//test/core/test_util:grpc_test_util",
+    ],
+)
